+++ fpdf/sign.py
@@ -1,4 +1,4 @@
-""""
+"""
 Module dedicated to document signature generation.
 
 The contents of this module are internal to fpdf2, and not part of the public API.