+++ fpdf/svg.py
@@ -1,4 +1,4 @@
-""""
+"""
 Utilities to parse SVG graphics into fpdf.drawing objects.
 
 The contents of this module are internal to fpdf2, and not part of the public API.