+++ fpdf/util.py
@@ -1,4 +1,4 @@
-""""
+"""
 Various utilities that could not be gathered logically in a specific module.
 
 The contents of this module are internal to fpdf2, and not part of the public API.